replacing audiofileplayer with just_audio plugin
