+++ example/lib/research_package_objects/linear_survey_objects.dart
@@ -222,7 +222,16 @@ final instructionStep = RPInstructionStep(
     text:
         "Please fill out this questionnaire!\n\nIn this questionnaire the questions will come after each other in a given order. You still have the chance to skip a some of them though.");
 
-final timerStep = RPTimerStep(
+final timerStep1 = RPTimerStep(
+  identifier: 'RPTimerStepID',
+  timeout: Duration(seconds: 5),
+  title:
+      "Please think for a moment about how your day was and note it down in the next step",
+  playSound: true,
+  autoSkip: true,
+);
+
+final timerStep2 = RPTimerStep(
   identifier: 'RPTimerStepID',
   timeout: Duration(seconds: 5),
   title:
@@ -235,8 +244,9 @@ final linearSurveyTask = RPOrderedTask(
   identifier: "surveyTaskID",
   steps: [
     instructionStep,
-    timerStep,
+    timerStep1,
     formStep,
+    timerStep2,
     // smokingQuestionStep,
     insulinStep,
     // wakeUpStep,
