+++ lib/src/loggers/activity_event_logger.dart
